requirements.txt
flask
pywebview
flask_socketio
